Assume untrusted when using read-only token for scanning.
Origin based PRs will be assumed to be trusted always. Fork PRs will be trusted
if the token has write access to origin and the fork PR author is a collaborator.

If the scanning API token does not have write access, then fork PRs will always
be assumed to be untrusted.